renaming with lower case the Exceptions module
