+++ grid2op/Exceptions/handlers_exceptions.py
@@ -6,7 +6,7 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
-from grid2op.Exceptions.ChronicsExceptions import ChronicsError
+from grid2op.Exceptions.grid2OpException import ChronicsError
 
 
 class HandlerError(ChronicsError):
